+++ source/Button/index.js
@@ -1,0 +1,6 @@
+/**
+ * @since 2016-09-05 13:21
+ * @author vivaxy
+ */
+
+export Button from './Button';